+++ src/index.js
@@ -1,3 +1,2 @@
 import './styles.scss';
 import './js/main';
-
