--- src/js/main/templates/articles.hbs
@@ -1,16 +1,0 @@
-{{#each this}}
- <li class="list-films__item">
-              <div class="foto-box">
-                <img
-                  src="{{}}"
-                  alt="{{}}"
-                  class="foto-img"
-                />
-              </div>
-              <p class="list-films__item-text">{{}}</p>
-              <div class="raiting-box">
-                <p class="list-films__item-info">{{}}</p>
-                <p class="list-films__item-rating">{{}}</p>
-              </div>
-            </li>
-{{/each}}
